complete login logic
