+++ public/index.html
@@ -25,7 +25,7 @@
   <title>React App</title>
 </head>
 
-<body>
+<body id="bigbomMarketplace">
   <noscript>
     You need to enable JavaScript to run this app.
   </noscript>
