+++ src/app/components/freelancer/index.js
@@ -2,6 +2,7 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { withStyles } from '@material-ui/core/styles';
 import { Link, Route, Switch } from 'react-router-dom';
+import { connect } from 'react-redux';
 
 import Dashboard from './Dashboard';
 import JobDetailBid from './JobDetailBid';
@@ -20,6 +21,12 @@ const styles = theme => ({
     },
 });
 class FreelancerContainer extends Component {
+    componentDidMount() {
+        const { isConnected, history } = this.props;
+        if (!isConnected) {
+            history.push('/login');
+        }
+    }
     render() {
         const { match } = this.props;
         const listSubLink = [
@@ -68,8 +75,23 @@ class FreelancerContainer extends Component {
 }
 
 FreelancerContainer.propTypes = {
+    history: PropTypes.object.isRequired,
+    isConnected: PropTypes.bool.isRequired,
     match: PropTypes.object.isRequired,
     classes: PropTypes.object.isRequired,
 };
 
-export default withStyles(styles)(FreelancerContainer);
+const mapStateToProps = state => {
+    return {
+        isConnected: state.homeReducer.isConnected,
+    };
+};
+
+const mapDispatchToProps = {};
+
+export default withStyles(styles)(
+    connect(
+        mapStateToProps,
+        mapDispatchToProps
+    )(FreelancerContainer)
+);
