+++ src/app/routers/index.js
@@ -2,17 +2,24 @@ import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import { Router, Route, Switch } from 'react-router-dom';
 import { Helmet } from 'react-helmet';
+import Eth from 'ethjs';
+import { connect } from 'react-redux';
 
 import asyncComponent from '../components/_asynComponent';
 import ScrollToTop from './scroll-to-top';
 import Header from '../containers/header';
 import Footer from '../containers/footer';
 import NotFound from '../components/NotFound';
+import Login from '../components/login';
 import RoutersUnAuthen from './RoutersUnAuthen';
 import RoutersAuthen from './RoutersAuthen';
 
+import Utils from '../_utils/utils';
+import { loginMetamask, logoutMetamask, setWeb3, setNetwork, setAccount } from '../components/home/actions';
+
 const Home = asyncComponent(() => import('../components/home'));
 
+let _this;
 class Routers extends PureComponent {
     constructor(props) {
         super(props);
@@ -20,16 +27,57 @@ class Routers extends PureComponent {
             isLogin: true,
             routes: RoutersUnAuthen,
         };
+        _this = this;
     }
 
     componentDidMount() {
         const { isLogin } = this.state;
+        const { setWeb3 } = this.props;
         if (isLogin) {
             this.setState({
                 routes: RoutersAuthen,
             });
         }
+        setWeb3(global.web3);
+        _this.checkMetamaskID = setInterval(() => {
+            _this.checkMetamask();
+        }, 1000);
     }
+
+    componentWillUnmount() {
+        clearInterval(this.checkMetamaskID);
+    }
+
+    static getDerivedStateFromProps(props, previousState) {
+        const { web3 } = props;
+        const state = {};
+        if (web3 && web3 !== previousState.web3) {
+            state.web3 = web3;
+            state.eth = new Eth(web3.currentProvider);
+        }
+        return state;
+    }
+
+    checkMetamask = async () => {
+        const { isConnected, logoutMetamask, setAccount, defaultAccount, history, setNetwork } = this.props;
+        const { web3 } = this.state;
+        if (isConnected) {
+            try {
+                const { account, network } = await Utils.connectMetaMask(web3);
+                if (defaultAccount !== account) {
+                    setAccount(account);
+                    setNetwork(network);
+                    if (defaultAccount) {
+                        logoutMetamask();
+                        console.log('logout');
+                        history.push('/login');
+                    }
+                }
+            } catch (error) {
+                logoutMetamask();
+            }
+        }
+    };
 
     render() {
         const { history } = this.props;
@@ -44,6 +92,7 @@ class Routers extends PureComponent {
                         <Header />
                         <Switch>
                             <Route exact path="/" component={Home} />
+                            <Route path="/login" component={Login} />
                             {routes.length && routes.map((route, key) => <Route key={key} {...route} />)}
                             <Route component={NotFound} />
                         </Switch>
@@ -56,7 +105,33 @@ class Routers extends PureComponent {
 }
 
 Routers.propTypes = {
+    loginMetamask: PropTypes.func.isRequired,
     history: PropTypes.object.isRequired,
+    isConnected: PropTypes.bool.isRequired,
+    defaultAccount: PropTypes.string.isRequired,
+    setWeb3: PropTypes.func.isRequired,
+    logoutMetamask: PropTypes.func.isRequired,
+    setAccount: PropTypes.func.isRequired,
+    setNetwork: PropTypes.func.isRequired,
 };
 
-export default Routers;
+const mapStateToProps = state => {
+    return {
+        web3: state.homeReducer.web3,
+        isConnected: state.homeReducer.isConnected,
+        defaultAccount: state.homeReducer.defaultAccount,
+    };
+};
+
+const mapDispatchToProps = {
+    loginMetamask,
+    setWeb3,
+    logoutMetamask,
+    setAccount,
+    setNetwork,
+};
+
+export default connect(
+    mapStateToProps,
+    mapDispatchToProps
+)(Routers);
