+++ src/styles/components/_home.scss
@@ -71,6 +71,10 @@
             width: 100%;
             padding: 40px 0;
         }
+        .note {
+            text-align: center;
+            width: 100%;
+        }
         .home-content-item {
             color: #444;
             .home-content-img {
